+++ src/GraphQL/loaders/defaultGraphQLMutations.js
@@ -2,14 +2,14 @@ import * as objectsMutations from './objectsMutations';
 import * as filesMutations from './filesMutations';
 import * as usersMutations from './usersMutations';
 import * as functionsMutations from './functionsMutations';
-import * as schemaMutations from './schemaMutations';
+import * as classSchemaMutations from './classSchemaMutations';
 
 const load = parseGraphQLSchema => {
   objectsMutations.load(parseGraphQLSchema);
   filesMutations.load(parseGraphQLSchema);
   usersMutations.load(parseGraphQLSchema);
   functionsMutations.load(parseGraphQLSchema);
-  schemaMutations.load(parseGraphQLSchema);
+  classSchemaMutations.load(parseGraphQLSchema);
 };
 
 export { load };
